Use relative import paths to avoid circular dependencies
