+++ jest.config.js
@@ -1,5 +1,7 @@
 module.exports = {
     preset: 'ts-jest',
+    moduleDirectories: ['node_modules', 'src'],
+    testPathIgnorePatterns: ['/node_modules/', 'build'],
     testEnvironment: 'jsdom',
     collectCoverageFrom: ['./src/**/*.ts'],
     setupFilesAfterEnv: ['jest-extended-snapshot'],
